+++ BasicPackage/ItemDomain/FoodDomain/Drag.cs
@@ -26,7 +26,7 @@ namespace InfinityWorldChess.ItemDomain.FoodDomain
 		public string ShowDescription => Description;
 
 		public IObjectAccessor<Sprite> ShowIcon => Icon;
-		public byte Score { get; set; }
+		public int Score { get; set; }
 
 		public int SaveIndex { get; set; }
